rgz/requirements.txt
aiogram==3.13.1
chromadb>=0.5.3
sentence-transformers==3.1.1
langchain-text-splitters==0.2.0
python-dotenv==1.0.1
pypdf==5.0.1
docx2txt==0.8
ollama==0.3.3

